input file for serial and parallel
this commit should be removed at certain point due to the fact that
input files are usually not committed with source code.

Because of the complexity of comparison between different experiments,
parameters change a lot. Simulation environments change.
It is better to keep track of input files and to respond well.

parallel script was added back in this patch.

Could be removed in the future